requirements.txt
asgiref==3.2.6
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
dj-rest-auth==1.0.0
Django==3.0.4
django-allauth==0.41.0
django-tastypie==0.14.3
djangorestframework==3.11.0
idna==2.9
oauthlib==3.1.0
python-dateutil==2.8.1
python-mimeparse==1.6.0
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
